requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
langgraph==0.0.27
langchain
python-dotenv
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-community==0.3.31
langchain-core==0.3.83
langchain-experimental==0.3.4
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
matplotlib==3.9.4
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-community==0.3.31
langchain-core==0.3.83
langchain-experimental==0.3.4
langchain-tavily==0.2.11
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
matplotlib==3.9.4
